Warehouse Order Tracking System/requirements.txt
streamlit
pandas
plotly
altair
bcrypt
python-dotenv
openpyxl
streamlit-option-menu
cryptography